feat: allow importing Decode/Encode directly
